requirements.txt
langchain==0.3.26
flask==3.1.1
sentence-transformers==4.1.0
pypdf==5.6.1
python-dotenv==1.1.0
langchain-pinecone==0.2.8
langchain-community==0.3.26

# Local embeddings (free)
transformers>=4.40.0
torch>=2.2.0

# Optional but recommended for local LLMs
ollama>=0.2.1

-e .
